emt-ai-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
vosk
imageio-ffmpeg
python-dotenv
